cmd-compress: Use multithreading for xz
xz compression is notoriously slow. Fortunately, it supports a
multi-threaded mode. The compression ratio is affected by this, but not
really in any significant way (see [1]).

On k8s though, we have to make sure to respect any cgroup limits set on
the CPU. (Though in CentOS right now, we run without any limits set at
all, so this will make a huge dent in our build times).

[1] https://yeah.nah.nz/misc/xz-thread/