Support overlay.d (and symlinks)
I'd like to encourage the pattern of adding e.g. `fedora-coreos-config`
as a git submodule, rather than forking it.  There are a lot
of subtle things going on here, and in practice no one wants to
"own" it.  If someone forks it even to do something as simple as add
packages, it's easy for things to diverge.

For the treefile, I think it works OK to just use `include: parent-config/blah.yaml`
in the derived treefile; we always expect people doing derivation to
make their own `manifest.yaml` anyways.

But I want to inherit from the overlay too.

My first idea here was to simply:
`ln -sr fedora-coreos-config/overlay overlay`
but that failed because (for probably bad reasons) ostree wants
a real directory as input.

Let's make this a bit more general by first adding support
for an `overlay.d` directory.  This allows us to "group"
overlays by purpose; if it helps you could think of them like
"inline packages".

Next, explicitly support symlinks here.  So for now for Silverblue
I'm just doing:
```
mkdir overlay.d
ln -sr fedora-coreos-config/overlay overlay.d/fcos
```

But perhaps in the future we'll change the FCOS `overlay` to `overlay.d`, and
this would allow us to cherry-pick from that in Silverblue and
other derived systems.