src/cmd-koji-upload: introduce file mutators
This is a rework of the file upload logic. One of the problems with Koji
Content Generators (cmd-koji-upload is such a thing) is that unless Koji
know about the content type being uploaded and the content type matches
a specific pattern than an upload will be rejected.

In the past we have uploaded uncompressed artifacts and renamed some
files to raw. A problem was hit when we tried to upload `vhd.gz` files.
We could have done the rename approach or come up with a more dynamic
solution.

This also addresses some wonky code. This refactor allow us to do the Koji
uploads, with out mucking with the rest of COSA by doing an upload
conversion for renaming true "raw" object like the kernel, and when a
supported format is not allowed to be uploaded in compressed fashion, do
a run-time decompress into a temporary directory.

This should allow almost _all_ uploads to be compressed while we purse
PR's like https://pagure.io/koji/pull-request/1608 to percolate.