cmd-fetch: add --write-lockfile-to option
`--update-lockfile` doesn't work for me because I run COSA in a container
and do a read-only bindmount of the src/config into the container. The
reason I've done this is so that I know the only changes I make to the
config are coming from me and not from what's being done inside the
container. This conflicts with `--update-lockfile` as it tries to write
the file in place. `--write-lockfile-to` allows one to specify where to
write the updated lockfile.