Add guestfs utility usage for s390x, to replace unsupported fw_cfg
In case of x86 fw_cfg is used to pass ignition config file to the KVM guest.
s390x does not support fw_cfg, so guestfs is option to use.
guestfs allows to modify qcow2 image before running it as guest.
Current implementation modifies not the base qcow2 image, but the temporary one, created based on the testing qcow2 image
Code has implementation of several calls to guestfish utility, it is expected that libguestfs-tools is installed on host

The list of commands to execute and their output for s390x case. Filesystems are not hardcoded in the code, but found based on their labels
  - guestfish --listen -a $IMG
         -- > GUESTFISH_PID=$PID; export GUESTFISH_PID
  - guestfish --remote -- run
  - guestfish --remote -- findfs-label boot
         -- >  /dev/sda1
  - guestfish --remote -- findfs-label root
         -- >  /dev/sda2
  - guestfish --remote -- mount /dev/sda2 /
  - guestfish --remote -- mount /dev/sda1 /boot
  - guestfish --remote -- mkdir -p /boot/ignition
  - guestfish --remote -- upload $HOST_IGNITION_FILE $GUEST_IGNITION_FILE
  - guestfish --remote -- umount-all
  - guestfish --remote -- exit