gcp: Add UEFI_COMPATIBLE and SECURE_BOOT
For FCOS we support both.  Mainly I'm adding these because
they seem to be prerequisites for accessing the "vTPM" which
is part of the "Shielded Cloud" umbrella.  We want access to the vTPM
to enable support for LUKS bound to vTPM for example.  Also,
for OpenShift, I think at some point we want to enable using TPM
as a "strong identity" for nodes.