build: Verify Python syntax in `make check`
We're not ready for the full flake8 (which honestly I find really
annoying anyways...) - but at least verify we're not shipping any
syntax errors.

Doing this since I saw a syntax error in an RHCOS build in
`cmd-buildextend-azure`, although I am not seeming to replicate
it locally from master at the moment.