cmdlib: don't extract repo tarball in tmp/repo/tmp
That repo is subject to cleanups by libostree. Just do it in the repo
directory itself so we still ensure that they're on the same device (and
so the `pull-local` can hardlink).

Note that `dir=f{repo}` just means to create the temporary directory in
which we extract there; we're not directly extracting on top of existing
OSTree repo.