platform/unprivqemu: Drop restrict=yes
Denying external network access indirectly breaks the `podman.go`
tests because podman uses bridging by default, expecting `/etc/resolv.conf`
to exist on the host and dying messily if it isn't.

I don't know of a reason for us to do this by default - probably
what we should do is something like only use `restrict=yes` for
tests that aren't flagged as requiring network.

But for now, let's make unprivqemu work like the other platforms
by default.  (As much as it can; this uses usermode TCP which has
weird limitations like no ICMP, and bugs of course, but it does
mostly work for what we need)