cmd-run: Allow `-p 0` to auto-choose SSH port
In fact, qemu does this for us already, but if we don't do it ourselves,
there'd be no way to report back to the user what port was chosen since
we exec into qemu. So they'd have to e.g. resort to `ss` to fish it out.

The `echo` there is semi-useful. It goes by really fast, though one can
always scroll back up to read it again. Really though, we expect people
using this to use `--ssh-config` (next patch).