qemu: turn on PDEATHSIG by default
To ensure that we don't leak the `qemu` process by accident, just always
set the `PDEATHSIG` attribute on it before spawning. That way, even if
`kola` is abruptly killed, `qemu` will also be killed.

The only exception is `kola spawn --remove=false`, where do *do* want to
leak the `qemu` process. Pass that down as a flag.

This also essentially enables support for the `--qemu-options` flag on
the `unprivqemu` platform.