buildprep/buildupload: Track "upstream" builds.json
Prep for "failed build" handling.  It is useful in general
to know what the difference is between the "upstream" and local
storage.  When we reun `buildprep`, record the URL as well as
the original `builds.json`.  When we `buildupload`, update our
cached copy.

(If we were using https://git-annex.branchable.com/ this type of
 stuff would come "for free")