guestfish: Support mounting RHCOS LUKS
We need the internal libguestfs bits to work; at least
`buildextend-installer` uses them.

First, the main mount code needs to detect this by looking for the
"luk_root"; if detected, use `luks-open` on it.

The thing that took me *FOREVER* to debug is that we need more
RAM for the libguestfs appliance, due to the new memory-hard argon2
KDF: https://gitlab.com/cryptsetup/cryptsetup/issues/372