installer: Simplify EFI extraction and hoist to a helper function
Prep for a PR to use *target* EFI binaries for main images, which should help
with RHCOS secure boot.

I also simplified the code a lot; I stared at it and just could not
see why it was listing the `EFI` directory and then checking out
subdirectories instead of just checking out the parent directory...directly.

This seems to result in the same filesystem layout with
`buildextend-live` as inspected with `guestfish`.