installer: Add volume ID and coreos.liveiso= karg referring to it
Prep for splitting out the rootfs.squashfs into a separate file
in the ISO mounted independently, rather than having it be
part of the initramfs.  The main problem I hit is that doing
FSBCOS live images didn't work as the resulting initramfs is
too big.

This patch just adds a volume ID (which means inserting the
ISO shows up with a nicer `/dev/disk/by-label/` instead of
just `CDROM`), and adds a kernel argument that can be used
by code in the initramfs to find that.

When this lands, we can change fedora-coreos-config to drop
its `coreos-iso` kernel argument.