cmd-buildupload: Set Content-Type and Content-Encoding headers
Automatically detect the type of data being uploaded and set the
`Content-{Type,Encoding}` headers accordingly if not overriden.

This allows one to do `curl --compressed` to decompress on the fly. Note
we don't automatically do the `Content-Disposition: inline; filename=`
trick; this was found to be more confusing than useful in practice.

Closes: https://github.com/coreos/fedora-coreos-tracker/issues/295