Drop cork and gangue
We don't use these for building RHCOS and FCOS. `coreos-assembler` is
the closest thing we have to `cork`, and `gangue` isn't really something
we need yet.

My other motivation for doing this is that this way, we can just do `cd
mantle && make` from the cosa Makefile (and similarly, `cd mantle &&
make install`) without having to say the specific binaries we want.