cmd-tag: Add --description flag
I want to be able to jot down more notes about a build when debugging.
This is super useful later on when looking back at the builds I have and
knowing which is which.

Add a `--description` flag for this. This essentiallly brings it even
closer to `git tag`, reviving the idea about just storing build metadata
in a git repo...  Anyway for now, let's build on top of what we have.