Change access mode to remove write permission from some artifacts
I was recently reading `man qemu` and noticed this snippet:

```
       -snapshot
           Write to temporary files instead of disk image files. In this case, the raw disk image you use is not written back. You can however force the write back by pressing C-a s.
```

`cosa run` uses `-snapshot` and I heavily rely on it not mutating
the disk image.  Qemu having this magic command to mutate the disk
image could lead to really evil debugging sessions.

And I verified that in fact it works today to press `C-a s`.

Further, we use `qemu-img create -b` with backing files, and
that will completely break if the underlying backing file is
mutated.

This patch closes that down by removing write access from our
generated disk images.  When running processes as an ordinary user
without `CAP_DAC_OVERRIDE`, this denies write access.

And I verified that qemu's `C-a s` does not mutate the underlying
disk image with this patch (it seems to silently ignore the failure).

This only covers a few artifacts, I didn't try to get them all yet.