oscontainer: Don't require a workdir
As of recently, the podman stack better supports being
invoked recursively.  Hence, it works to just
`podman unshare cosa oscontainer build ...` inside an existing
container for me.

We can hence now just write to the primary container storage, if
we want to - and I think doing so is more ergonomic for development/testing
at least.  And we could consider doing this in prod builds as well;
conceptually the build user's homedir is transient anyways.