run: Add --ipv6only
We'd like to support IPv6 only environments; make it convenient
to do that with `cosa run`, also prep for testing a kola patch.

However, I can't get `ip=dhcp6` to work with this, seems like
nothing is responding on the qemu side.  Learned a bit about
slirp, but haven't debugged it yet.