buildextend-rojig: Support unprivileged mode
Trying to actually deploy this in the FSBCOS pipeline, it failed
because it wants privileges, just like the main compose path.

Tweak things so that we do the same "run in VM if not privileged"
that main composes do.

The code/logic for this stuff really needs cleaning up and deduping,
it's quite messy.  Particularly because the supermin API is only
in shell.

We had some duplicates between `deps.txt` and `vmdeps.txt`; let's
consistently have the rpm-ostree bits in `vmdeps.txt`.  We
specifically need `rpm-build` in `vmdeps.txt` for rojig.