build-sys: Unify deps/vmdeps, clean up dependency bits
We were missing `rpm-build` in vmdeps which is needed for rojig.
Change things so that vmdeps are *also* dependencies of the main
container build, rather than needing to specify in both places.

Rework the bits around dependencies so we have a simple shell
script that is called by both the Dockerfile and during
runtime.  Having the Dockerfile source `cmdlib.sh` was a confusing
hack prone to breakage, since `cmdlib.sh` is all oriented around
being used at runtime.

No good reason to `depcheck` at build time really anyways, that's
really a testing thing.  As soon as we run `cosa` it'll depcheck.

Also, this reveals that `bind-export-libs` needs to be dropped
since it no longer exists in F31.  `supermin` apparently ignores
nonexistent packages.