ore/aws/copy-image: stream results back to stdout
Rather than accumulate all the results and then dump it all on stdout,
stream the output as line-separated JSON objects as the goroutines
complete.

This is technically a compatibility break since we change the format of
the output. However, I'm not aware of any clients which use `copy-image`
directly and not through cosa.

Prep for making use of this in cosa to speed up `cosa aws-replicate`.